% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dv_update_metadata}
\alias{dv_update_metadata}
\title{Update metadata from upstream}
\usage{
dv_update_metadata(
  url = paste0("https://raw.githubusercontent.com/google/libphonenumber/master/",
    "resources/PhoneNumberMetadata.xml")
)
}
\arguments{
\item{url}{URL to the raw PhoneNumberMetadata.xml file. Defaults to the
master branch on GitHub.}
}
\value{
Invisibly returns the path to the cached metadata file.
}
\description{
Downloads the latest \code{PhoneNumberMetadata.xml} from Google's libphonenumber
repository, parses it, and saves the result to the user's cache directory.
The updated metadata is used for the remainder of the session.
}
\details{
Requires the \pkg{xml2} package.
}
\examples{
\donttest{
dv_update_metadata()
}

}
